integrate egg parser into importer
