+++ app/Services/Eggs/Sharing/EggImporterService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services\Eggs\Sharing;
 
+use App\Exceptions\Service\InvalidFileUploadException;
 use Ramsey\Uuid\Uuid;
 use Illuminate\Support\Arr;
 use App\Models\Egg;
@@ -14,7 +15,18 @@ use Spatie\TemporaryDirectory\TemporaryDirectory;
 
 class EggImporterService
 {
-    public function __construct(protected ConnectionInterface $connection, protected EggParserService $parser)
+    public const UPGRADE_VARIABLES = [
+        'server.build.env.SERVER_IP' => 'server.allocations.default.ip',
+        'server.build.default.ip' => 'server.allocations.default.ip',
+        'server.build.env.SERVER_PORT' => 'server.allocations.default.port',
+        'server.build.default.port' => 'server.allocations.default.port',
+        'server.build.env.SERVER_MEMORY' => 'server.build.memory_limit',
+        'server.build.memory' => 'server.build.memory_limit',
+        'server.build.env.' => 'server.environment.',
+        'server.build.environment.' => 'server.environment.',
+    ];
+
+    public function __construct(protected ConnectionInterface $connection)
     {
     }
 
@@ -25,7 +37,7 @@ class EggImporterService
      */
     public function fromFile(UploadedFile $file, Egg $egg = null): Egg
     {
-        $parsed = $this->parser->handle($file);
+        $parsed = $this->parseFile($file);
 
         return $this->connection->transaction(function () use ($egg, $parsed) {
             $uuid = $parsed['uuid'] ?? Uuid::uuid4()->toString();
@@ -37,7 +49,7 @@ class EggImporterService
                 'copy_script_from' => null,
             ]);
 
-            $egg = $this->parser->fillFromParsed($egg, $parsed);
+            $egg = $this->fillFromParsed($egg, $parsed);
             $egg->save();
 
             // Update existing variables or create new ones.
@@ -71,5 +83,90 @@ class EggImporterService
         file_put_contents($tmpPath, file_get_contents($url));
 
         return $this->fromFile(new UploadedFile($tmpPath, $info['basename'], 'application/json'), $egg);
+    }
+
+    /**
+     * Takes an uploaded file and parses out the egg configuration from within.
+     *
+     * @throws \JsonException
+     * @throws \App\Exceptions\Service\InvalidFileUploadException
+     */
+    protected function parseFile(UploadedFile $file): array
+    {
+        if ($file->getError() !== UPLOAD_ERR_OK) {
+            throw new InvalidFileUploadException('The selected file was not uploaded successfully');
+        }
+
+        $parsed = json_decode($file->getContent(), true, 512, JSON_THROW_ON_ERROR);
+
+        $version = $parsed['meta']['version'] ?? '';
+
+        $parsed = match ($version) {
+            'PTDL_v1' => $this->convertToV2($parsed),
+            'PTDL_v2' => $parsed,
+            default => throw new InvalidFileUploadException('The JSON file provided is not in a format that can be recognized.')
+        };
+
+        // Make sure we only use recent variable format from now on
+        $parsed['config']['files'] = str_replace(
+            array_keys(self::UPGRADE_VARIABLES),
+            array_values(self::UPGRADE_VARIABLES),
+            $parsed['config']['files'] ?? '',
+        );
+
+        return $parsed;
+    }
+
+    /**
+     * Fills the provided model with the parsed JSON data.
+     */
+    protected function fillFromParsed(Egg $model, array $parsed): Egg
+    {
+        return $model->forceFill([
+            'name' => Arr::get($parsed, 'name'),
+            'description' => Arr::get($parsed, 'description'),
+            'features' => Arr::get($parsed, 'features'),
+            'docker_images' => Arr::get($parsed, 'docker_images'),
+            'file_denylist' => Collection::make(Arr::get($parsed, 'file_denylist'))
+                ->filter(fn ($value) => !empty($value)),
+            'update_url' => Arr::get($parsed, 'meta.update_url'),
+            'config_files' => Arr::get($parsed, 'config.files'),
+            'config_startup' => Arr::get($parsed, 'config.startup'),
+            'config_logs' => Arr::get($parsed, 'config.logs'),
+            'config_stop' => Arr::get($parsed, 'config.stop'),
+            'startup' => Arr::get($parsed, 'startup'),
+            'script_install' => Arr::get($parsed, 'scripts.installation.script'),
+            'script_entry' => Arr::get($parsed, 'scripts.installation.entrypoint'),
+            'script_container' => Arr::get($parsed, 'scripts.installation.container'),
+        ]);
+    }
+
+    /**
+     * Converts a PTDL_V1 egg into the expected PTDL_V2 egg format. This just handles
+     * the "docker_images" field potentially not being present, and not being in the
+     * expected "key => value" format.
+     */
+    protected function convertToV2(array $parsed): array
+    {
+        // Maintain backwards compatability for eggs that are still using the old single image
+        // string format. New eggs can provide an array of Docker images that can be used.
+        if (!isset($parsed['images'])) {
+            $images = [Arr::get($parsed, 'image') ?? 'nil'];
+        } else {
+            $images = $parsed['images'];
+        }
+
+        unset($parsed['images'], $parsed['image']);
+
+        $parsed['docker_images'] = [];
+        foreach ($images as $image) {
+            $parsed['docker_images'][$image] = $image;
+        }
+
+        $parsed['variables'] = array_map(function ($value) {
+            return array_merge($value, ['field_type' => 'text']);
+        }, $parsed['variables']);
+
+        return $parsed;
     }
 }
